// src/tools/Interfaces/OtherInterface.ts
export interface SettingStyleArena {
    border: boolean
    infoIndex: boolean
    infoCountPixel: boolean
    sizeCell: number
    infoLvl: boolean
    colorBackgroundArena: string
}

export interface LocalStatisticsInfo {
    countGames: number
    totalTime: number
    totalSteps: number
    countWinTeamOne: number
    countWinTeamTwo: number
}